src/test/java/ru/job4j/loop/Task152Test.java
package ru.job4j.loop;

import org.junit.jupiter.api.Test;

import java.io.ByteArrayOutputStream;
import java.io.PrintStream;
import java.util.StringJoiner;

import static org.assertj.core.api.Assertions.assertThat;

class Task152Test {
    @Test
    public void whenTest3Rows() {
        ByteArrayOutputStream out = new ByteArrayOutputStream();
        System.setOut(new PrintStream(out));
        Task152.loop();
        StringJoiner expected = new StringJoiner(
                System.lineSeparator(), "", System.lineSeparator())
                .add("10 12 14 16 18 20").add("12 15 18")
                .add("12 16 20");
        assertThat(out.toString()).isEqualTo(expected.toString());
    }
}